+++ .gitignore
@@ -7,3 +7,4 @@ cache/
 htmlcov/
 
 cdk.out/*
+build/*
