+++ deploy/cdk.py
@@ -2,7 +2,6 @@
 from aws_cdk import core
 
 from lib.lambdas import Lambdas
-from lib.layers import Layers
 from lib.utils import clean_pycache
 
 clean_pycache()
@@ -11,7 +10,6 @@ app = core.App()
 
 env = {"region": "eu-west-1"}
 
-layers_stack = Layers(app, "anime-layers", env=env)
-Lambdas(app, "anime-lambdas", layers_stack.layers, env=env)
+Lambdas(app, "anime-lambdas", env=env)
 
 app.synth()
