--- deploy/lib/layers.py
@@ -1,42 +1,0 @@
-import os
-import shutil
-import subprocess
-
-from aws_cdk import core
-from aws_cdk.aws_lambda import LayerVersion, Code, Runtime
-
-CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
-LAYERS_DIR = os.path.join(CURRENT_DIR, "..", "..", "src", "layers")
-BUILD_FOLDER = os.path.join(CURRENT_DIR, "..", "..", "build")
-
-
-class Layers(core.Stack):
-
-    def __init__(self, app: core.App, id: str, **kwargs) -> None:
-        super().__init__(app, id, **kwargs)
-        self.layers = {}
-        self._create_layers()
-
-    def _create_layers(self):
-        if os.path.isdir(BUILD_FOLDER):
-            shutil.rmtree(BUILD_FOLDER)
-        os.mkdir(BUILD_FOLDER)
-
-        for layer in os.listdir(LAYERS_DIR):
-            layer_folder = os.path.join(LAYERS_DIR, layer)
-            build_folder = os.path.join(BUILD_FOLDER, layer)
-            shutil.copytree(layer_folder, build_folder)
-
-            requirements_path = os.path.join(build_folder, "requirements.txt")
-
-            if os.path.isfile(requirements_path):
-                packages_folder = os.path.join(build_folder, "python", "lib", "python3.8", "site-packages")
-                print(f"Installing layer requirements to target: {os.path.abspath(packages_folder)}")
-                subprocess.check_output(["pip", "install", "-r", requirements_path, "-t", packages_folder])
-
-            self.layers[layer] = LayerVersion(
-                self,
-                layer,
-                code=Code.from_asset(path=build_folder),
-                compatible_runtimes=[Runtime.PYTHON_3_8],
-            )
